+++ application/views/fleet/fleetindexx.php
@@ -1,6 +1,7 @@
 <div class="panel panel-default">
 	<div class="panel-heading">{pagetitle}</div>
 	<div class="panel-body">
+    {add}
 		<table class="table table-striped table-hover">
 
 		<!-- table heading -->
@@ -8,6 +9,7 @@
         <th>Plane ID</th>
         <th>Manufacturer</th>
         <th>Model</th>
+        <th>Action<th>
       </tr>
 
       	<!-- table content, displaying info of planes -->
@@ -16,6 +18,7 @@
         <td><a href="/fleet/edit/{id}"><u>{id}</u></a></td>
         <td>{manufacturer}</td>
         <td>{model}</td>
+        <td><a href="fleet/delete/{id}"><button class="btn btn-danger">Delete</button></a><a href="fleet/edit/{id}"><button class="btn btn-info">Edit</button></a></td>
       </tr>
       {/planes}